Separate literal_value_extractor
