+++ zetasql/fuzzing/component/fuzz_targets/prepared_expression_target.h
@@ -28,11 +28,9 @@ class PreparedExpressionTarget : public FuzzTarget {
   void Execute() override;
 
  private:
-  std::unique_ptr<std::string> sql_expression;
-  std::unique_ptr<zetasql::ParameterValueMap> columns =
-      std::make_unique<zetasql::ParameterValueMap>();
-  std::unique_ptr<zetasql::ParameterValueMap> parameters =
-      std::make_unique<zetasql::ParameterValueMap>();
+  std::unique_ptr<std::string> sql_expression_;
+  std::unique_ptr<zetasql::ParameterValueMap> columns_;
+  std::unique_ptr<zetasql::ParameterValueMap> parameters_;
 };
 
 }  // namespace zetasql_fuzzer
